Introduce `whoami` flag to the `login` command
Scripts need a clean and consistent way to query the Firebase CLI for the current user's auth state and email address. Common practice is to parse the output from `firebase login` to get this information. This extra step adds unnecessary overhead to scripts, to the developers who write the scripts, and to the systems that run those scripts. This may be a small matter in an instance, but it's an additive issue that can be addressed elegantly and simply. By introducing a flag, like `whoami` or something similar, we can improve the development experience, thereby encouraging continued use of the official CLI.